python/requirements.txt
cffi==1.15.0
click==8.1.3
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycmarkgfm==1.1.0
pycparser==2.21
PyYAML==6.0
Werkzeug==2.1.2
